refactor: 메뉴 상품과 상품 같은 패키지로 이동
