+++ kitchenpos-core/src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.application;
+package kitchenpos.product.application;
 
 import kitchenpos.common.dto.request.CreateProductRequest;
-import kitchenpos.menu.domain.Product;
-import kitchenpos.menu.domain.repository.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
