+++ kitchenpos-core/src/main/java/kitchenpos/product/domain/repository/MenuProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.domain.repository;
+package kitchenpos.product.domain.repository;
 
-import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.product.domain.MenuProduct;
 
 import java.util.List;
 import java.util.Optional;
